v2/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.0
pgvector==0.3.6
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.12
anthropic==0.40.0
openai==1.55.0
google-genai==1.14.0
httpx==0.27.0
PyMuPDF==1.24.0
python-docx==1.1.0
markdown==3.7
